solved some errors related to the post files

+++ social_media/src/main/java/com/example/social_media/post/PostController.java
@@ -47,7 +47,7 @@ public class PostController {
     }
 
     @GetMapping("/user/{userId}")
-    public ResponseEntity<List<PostResponse>> getUserPosts(@PathVariable Integer userId) {
+    public ResponseEntity<List<PostResponse>> getUserPosts(@PathVariable Long userId) {
         User user = userService.findById(userId);
         List<Post> posts = postService.findByUser(user);
         List<PostResponse> responses = posts.stream()
