+++ social_media/src/main/java/com/example/social_media/post/PostRepository.java
@@ -18,5 +18,5 @@ public interface PostRepository extends JpaRepository<Post, Long> {
     @Query("SELECT p FROM Post p LEFT JOIN FETCH p.comments WHERE p.id = :id")
     Post findByIdWithComments(Long id);
     
-    boolean existsByIdAndUserId(Long postId, Integer userId);
+    boolean existsByIdAndUserId(Long postId, Long userId);
 }
